requirements.txt
google-genai>=0.7.0
streamlit>=1.35.0
yt-dlp>=2024.12.0
faster-whisper>=1.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pymupdf>=1.24.0
beautifulsoup4>=4.12.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
streamlit-extras>=0.4.0
